# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LLMNightRun",
    "description": "멀티 LLM 통합 자동화 플랫폼",
    "version": "0.1.0",
    "x-logo": {
      "url": "/static/logo.png",
      "altText": "LLMNightRun"
    }
  },
  "paths": {
...
    "/github/config": {
      "get": {
        "tags": [
          "github"
        ],
        "summary": "Get Github Config",
        "description": "GitHub 설정 조회",
        "operationId": "get_github_config_github_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "github"
        ],
        "summary": "Update Github Config",
        "description": "GitHub 설정 업데이트",
        "operationId": "update_github_config_github_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubConfig"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
...
    "/docs-manager/list": {
      "get": {
        "tags": [
          "docs-manager"
        ],
        "summary": "List Documents",
        "description": "사용 가능한 문서 목록과 상태 조회",
        "operationId": "list_documents_docs_manager_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Documents Docs Manager List Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/docs-manager/generate": {
      "post": {
        "tags": [
          "docs-manager"
        ],
        "summary": "Generate Documents",
        "description": "선택한 문서 생성 또는 업데이트",
        "operationId": "generate_documents_docs_manager_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Generate Documents Docs Manager Generate Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/docs-manager/content/{doc_type}": {
      "get": {
        "tags": [
          "docs-manager"
        ],
        "summary": "Get Document Content",
        "description": "특정 문서의 내용 조회",
        "operationId": "get_document_content_docs_manager_content__doc_type__get",
        "parameters": [
          {
            "name": "doc_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "문서 유형",
              "title": "Doc Type"
            },
            "description": "문서 유형"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Document Content Docs Manager Content  Doc Type  Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/docs-manager/github/config": {
      "get": {
        "tags": [
          "docs-manager"
        ],
        "summary": "Get Github Config",
        "description": "GitHub 연동 설정 조회",
        "operationId": "get_github_config_docs_manager_github_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Github Config Docs Manager Github Config Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "docs-manager"
        ],
        "summary": "Set Github Config",
        "description": "GitHub 연동 설정 저장",
        "operationId": "set_github_config_docs_manager_github_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GitHubConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Set Github Config Docs Manager Github Config Post"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
    "/docs-manager/github/status": {
      "get": {
        "tags": [
          "docs-manager"
        ],
        "summary": "Get Github Status",
        "description": "GitHub 리포지토리 상태 조회",
        "operationId": "get_github_status_docs_manager_github_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Github Status Docs Manager Github Status Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "validation_error",
                  "message": "데이터 검증 오류가 발생했습니다",
                  "detail": [
                    {
                      "loc": [
                        "body",
                        "name"
                      ],
                      "msg": "필수 항목입니다"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않음",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "unauthorized",
                  "message": "API 키가 필요합니다",
                  "detail": null
                }
              }
            }
          },
          "500": {
            "description": "서버 내부 오류",
            "content": {
              "application/json": {
                "example": {
                  "error_code": "internal_server_error",
                  "message": "서버 내부 오류가 발생했습니다",
                  "detail": "오류에 대한 자세한 내용"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentGenerateRequest": {
        "properties": {
          "doc_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Doc Types"
          },
          "force_update": {
            "type": "boolean",
            "title": "Force Update",
            "default": false
          },
          "push_to_github": {
            "type": "boolean",
            "title": "Push To Github",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "doc_types"
        ],
        "title": "DocumentGenerateRequest"
      },
...
      "GitHubConfig": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "default": "github_config"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "default_branch": {
            "type": "string",
            "title": "Default Branch",
            "default": "main"
          },
          "auto_commit": {
            "type": "boolean",
            "title": "Auto Commit",
            "default": false
          },
          "docs_directory": {
            "type": "string",
            "title": "Docs Directory",
            "default": "docs"
          },
          "commit_message_template": {
            "type": "string",
            "title": "Commit Message Template",
            "default": "docs: Update {doc_type} documentation"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "title": "GitHubConfig",
        "description": "GitHub 설정 모델"
      },
      "GitHubConfigRequest": {
        "properties": {
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "default": "main"
          },
          "auth_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth Token"
          },
          "auto_push": {
            "type": "boolean",
            "title": "Auto Push",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "repo_url"
        ],
        "title": "GitHubConfigRequest"
      },
      "GitHubConfigUpdateRequest": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "repo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repo Url"
          },
          "token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "default_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Branch"
          },
          "auto_commit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Commit"
          },
          "docs_directory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docs Directory"
          },
          "commit_message_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Message Template"
          }
        },
        "type": "object",
        "title": "GitHubConfigUpdateRequest",
        "description": "GitHub 설정 업데이트 요청 모델"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
  "servers": [
    {
      "url": "",
      "description": "개발 서버"
    }
  ]
}